requirements.txt
pandas>=2.0.0
numpy>=1.24.0
fairlearn>=0.10.0
scikit-learn>=1.3.0
google-genai>=0.1.0
python-dotenv>=1.0.0